Moved cfc and traverse into the builder module Moved SchemaContext into builder/types Moved FactAddress into memory/interface Add cfc as a field in IRuntime and use this when it's accessible (in doc, cell, and storage)
