+++ packages/builder/src/cfc.ts
