+++ packages/builder/src/index.ts
@@ -60,6 +60,7 @@ export {
   type OpaqueRefMethods,
   type Recipe,
   type RecipeFactory,
+  type SchemaContext,
   type Static,
   type StreamAlias,
   type toJSON,
@@ -73,7 +74,7 @@ export {
 } from "./types.ts";
 export { type Schema, schema } from "./schema-to-ts.ts";
 export { AuthSchema } from "./schema-lib.ts";
-
+export { Classification, ContextualFlowControl } from "./cfc.ts";
 // This should be a separate package, but for now it's easier to keep it here.
 export {
   createJsonSchema,
