+++ packages/builder/src/opaque-ref.ts
@@ -15,7 +15,7 @@ import { hasValueAtPath, setValueAtPath } from "./utils.ts";
 import { getTopFrame, recipe } from "./recipe.ts";
 import { createNodeFactory } from "./module.ts";
 import { SchemaWithoutCell } from "./schema-to-ts.ts";
-import { ContextualFlowControl } from "../../runner/src/index.ts";
+import { ContextualFlowControl } from "./cfc.ts";
 
 let mapFactory: NodeFactory<any, any>;
