+++ packages/memory/consumer.ts
@@ -31,7 +31,6 @@ import {
   Reference,
   Result,
   Revision,
-  SchemaContext,
   SchemaNone,
   SchemaPathSelector,
   SchemaQuery,
@@ -48,7 +47,7 @@ import {
   UCAN,
   UTCUnixTimestampInSeconds,
 } from "./interface.ts";
-import type { JSONValue } from "@commontools/builder";
+import type { JSONValue, SchemaContext } from "@commontools/builder";
 import { refer } from "./reference.ts";
 import * as Socket from "./socket.ts";
 import {
