+++ packages/runner/src/cell.ts
@@ -4,6 +4,7 @@ import {
   ID,
   ID_FIELD,
   type JSONSchema,
+  type Schema,
 } from "@commontools/builder";
 import { type DeepKeyLookup, type DocImpl, isDoc } from "./doc.ts";
 import { getEntityId } from "./doc-map.ts";
@@ -20,8 +21,6 @@ import type { ReactivityLog } from "./scheduler.ts";
 import { type EntityId } from "./doc-map.ts";
 import { type Cancel, isCancel, useCancelGroup } from "./cancel.ts";
 import { validateAndTransform } from "./schema.ts";
-import { type Schema } from "@commontools/builder";
-import { ContextualFlowControl } from "./index.ts";
 
 /**
  * This is the regular Cell interface, generated by DocImpl.asCell().
@@ -290,7 +289,6 @@ function createRegularCell<T>(
   schema?: JSONSchema,
   rootSchema?: JSONSchema,
 ): Cell<T> {
-  const cfc = new ContextualFlowControl();
   const self = {
     get: () => validateAndTransform(doc, path, schema, log, rootSchema),
     set: (newValue: Cellify<T>) =>
@@ -381,7 +379,7 @@ function createRegularCell<T>(
     key: <K extends T extends Cell<infer S> ? keyof S : keyof T>(
       valueKey: K,
     ): T extends Cell<infer S> ? Cell<S[K & keyof S]> : Cell<T[K]> => {
-      const childSchema = cfc.getSchemaAtPath(
+      const childSchema = doc.runtime.cfc.getSchemaAtPath(
         schema,
         [valueKey.toString()],
         rootSchema,
