+++ packages/runner/src/index.ts
@@ -32,5 +32,4 @@ export {
   followAliases,
   maybeGetCellLink,
 } from "./utils.ts";
-export { Classification, ContextualFlowControl } from "./cfc.ts";
 export * from "./recipe-manager.ts";
