+++ packages/runner/src/storage/base.ts
@@ -2,7 +2,7 @@ import type { Cancel } from "../cancel.ts";
 import type { EntityId } from "../doc-map.ts";
 import type { Entity, Result, Unit } from "@commontools/memory/interface";
 import { log } from "../log.ts";
-import { SchemaContext } from "@commontools/memory/interface";
+import { SchemaContext } from "@commontools/builder";
 
 export type { Result, Unit };
